tests are fixed
Former-commit-id: d32f66cd39a12794e475582016dce7ea99979f95

+++ src/Tests/WB.Tests.Integration/FileChunkStorageTests.cs
@@ -12,13 +12,15 @@ namespace WB.Tests.Integration
     [TestFixture]
     public class FileChunkStorageTests
     {
-        private const string FolderPath = "/";
+        private const string FolderPath = ".";
         private const string FolderName = "SyncData";
 
         [SetUp]
         public void SetUp()
         {
-            Directory.Delete(Path.Combine(FolderPath, FolderName), true);
+            var dirPath = Path.Combine(FolderPath, FolderName);
+            if (Directory.Exists(dirPath))
+                Directory.Delete(Path.Combine(FolderPath, FolderName), true);
         }
 
         [Test]
